+++ packages/account/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prosopo/account
 
+## 2.8.82
+### Patch Changes
+
+- Updated dependencies [4b1cb19]
+  - @prosopo/types@5.4.0
+  - @prosopo/common@3.1.52
+  - @prosopo/fingerprint@2.7.36
+  - @prosopo/keyring@2.9.79
+
+## 2.8.81
+### Patch Changes
+
+- Updated dependencies [b30ad41]
+  - @prosopo/types@5.3.0
+  - @prosopo/fingerprint@2.7.35
+  - @prosopo/keyring@2.9.78
+
 ## 2.8.80
 ### Patch Changes