+++ packages/api-express-router/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @prosopo/api-express-router
 
+## 3.1.75
+### Patch Changes
+
+  - @prosopo/env@3.6.44
+
+## 3.1.74
+### Patch Changes
+
+- Updated dependencies [4b1cb19]
+  - @prosopo/types@5.4.0
+  - @prosopo/locale@3.4.0
+  - @prosopo/common@3.1.52
+  - @prosopo/env@3.6.43
+
+## 3.1.73
+### Patch Changes
+
+- Updated dependencies [b30ad41]
+  - @prosopo/types@5.3.0
+  - @prosopo/env@3.6.42
+
 ## 3.1.72
 ### Patch Changes