+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @prosopo/cli
 
+## 3.7.27
+### Patch Changes
+
+  - @prosopo/env@3.6.44
+  - @prosopo/provider@5.5.1
+
+## 3.7.26
+### Patch Changes
+
+- Updated dependencies [4b1cb19]
+  - @prosopo/types@5.4.0
+  - @prosopo/locale@3.4.0
+  - @prosopo/api@4.1.0
+  - @prosopo/provider@5.5.0
+  - @prosopo/common@3.1.52
+  - @prosopo/env@3.6.43
+  - @prosopo/keyring@2.9.79
+
+## 3.7.25
+### Patch Changes
+
+- Updated dependencies [b30ad41]
+  - @prosopo/types@5.3.0
+  - @prosopo/provider@5.4.0
+  - @prosopo/api@4.0.15
+  - @prosopo/env@3.6.42
+  - @prosopo/keyring@2.9.78
+
 ## 3.7.24
 ### Patch Changes