+++ packages/detector/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/detector
 
+## 3.5.31
+### Patch Changes
+
+- Updated dependencies [4b1cb19]
+  - @prosopo/types@5.4.0
+
+## 3.5.30
+### Patch Changes
+
+- Updated dependencies [b30ad41]
+  - @prosopo/types@5.3.0
+
 ## 3.5.29
 ### Patch Changes