+++ packages/ipinfo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prosopo/ipinfo
 
+## 0.3.17
+### Patch Changes
+
+- Updated dependencies [4b1cb19]
+  - @prosopo/types@5.4.0
+
+## 0.3.16
+### Patch Changes
+
+- Updated dependencies [b30ad41]
+  - @prosopo/types@5.3.0
+
 ## 0.3.15
 ### Patch Changes