+++ packages/procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @prosopo/procaptcha-wrapper
 
+## 2.6.121
+### Patch Changes
+
+- Updated dependencies [4b1cb19]
+  - @prosopo/types@5.4.0
+  - @prosopo/locale@3.4.0
+
+## 2.6.120
+### Patch Changes
+
+- Updated dependencies [b30ad41]
+  - @prosopo/types@5.3.0
+
 ## 2.6.119
 ### Patch Changes